+++ app/src/main/java/com/pvsagar/smartlockscreen/EditEnvironment.java
@@ -697,7 +697,7 @@ public class EditEnvironment extends ActionBarActivity {
                     return;
                 }
 
-                environmentVariables.add(new LocationEnvironmentVariable((float)latLocation,(float)lonLocation,(int)radLocation,locationName));
+                environmentVariables.add(new LocationEnvironmentVariable(latLocation,lonLocation,(int)radLocation,locationName));
             }
 
             if(!bluetoothFlag && !wifiFlag && !locationFlag){
